Move TaskManager ray actor class into its own file
